Fixing legend url with custom format.
The tests didn't show the issue because of the use of OpenLayers.Util.isEquivalentUrl, but the way the addition of the default image/gif format is implemented, it will be appended to a format that was defined in baseParams. To avoid this, we append to the url after baseParams were appended.